Fix merge hell and eslint issues
