+++ backend/index.js
@@ -3,17 +3,18 @@ const app = express();
 const PORT = process.env.PORT || 4000;
 const cors = require("cors");
 
-//middleware
+// middleware
 app.use(express.json());
 app.use(cors());
 
 // routes
 
 // get
-app.get("/", async (req, res) => {
+app.get("/", (req, res) => {
   res.send("Hello World!");
 });
 
 app.listen(PORT, () => {
+  // TODO: Replace with proper logger
   console.log(`Server is running in port: ${PORT}`);
 });
