+++ frontend/src/components/Carousel/Carousel.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import Slider from "react-slick";
-import styles from "./Carousel.module.css"; // add css
+
+import styles from "./Carousel.module.css";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 
@@ -10,30 +11,25 @@ export default function SimpleSlider() {
     infinite: true,
     speed: 500,
     slidesToShow: 1,
-    slidesToScroll: 1,
+    slidesToScroll: 1
   };
   return (
     <div>
-            <p className={styles.label}>Get your next
-            </p>
-    <Slider {...settings}>
-      <div>
-      <p className={`${styles.slideText} ${styles.Text1}`}>
-        weeknight dinner idea
-        </p>
-      </div>
-      <div>
-      <p className={`${styles.slideText} ${styles.Text2}`}>home décor idea
-        </p>
-      </div>
-      <div>
-      <p className={`${styles.slideText} ${styles.Text3}`}>new outfit</p>
-      </div>
-      <div>
-      <p className={`${styles.slideText} ${styles.Text4}`}>green thumb idea
-        </p>
-      </div>
-    </Slider>
+      <p className={styles.label}>Get your next</p>
+      <Slider {...settings}>
+        <div>
+          <p className={`${styles.slideText} ${styles.Text1}`}>weeknight dinner idea</p>
+        </div>
+        <div>
+          <p className={`${styles.slideText} ${styles.Text2}`}>home décor idea</p>
+        </div>
+        <div>
+          <p className={`${styles.slideText} ${styles.Text3}`}>new outfit</p>
+        </div>
+        <div>
+          <p className={`${styles.slideText} ${styles.Text4}`}>green thumb idea</p>
+        </div>
+      </Slider>
     </div>
   );
 }
