+++ frontend/src/components/ExampleComponent.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 export default function ExampleComponent() {
   return <div>Example Component</div>;
