+++ frontend/src/components/ExplorePageSection1/CardDetailPage.jsx
@@ -1,8 +1,9 @@
 import React from "react";
 import { useParams } from "react-router-dom";
+
+import Breadcrumb from "./Breadcrumb";
 import styles from "./CardDetailPage.module.css";
 import { cards } from "./cardDetails";
-import Breadcrumb from "./Breadcrumb";
 
 const CardDetailPage = () => {
   const { id } = useParams();
