+++ frontend/src/components/Forms/Login/Login.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { CiMail } from "react-icons/ci";
-import { CiLock } from "react-icons/ci";
+import { CiMail, CiLock } from "react-icons/ci";
 import { FaChevronDown } from "react-icons/fa";
+
 import styles from "./Login.module.css";
 
 const Login = () => (
