+++ frontend/src/components/Navbar/Navbar.jsx
@@ -1,7 +1,9 @@
 import React from "react";
+
 import Header from "../Header/Header";
-import style from "./navbar.module.css";
 import Modal from "../Modal/Modal";
+
+import style from "./navbar.module.css";
 
 export default function Navbar() {
   return (
