+++ frontend/src/components/OptionsList/OptionsList.jsx
@@ -1,4 +1,5 @@
 import { useState, useRef, useEffect } from "react";
+
 import styles from "./OptionsList.module.css";
 import OptionsListItem from "./OptionsListItem";
