+++ frontend/src/components/PinDetailComponent.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+
 import styles from "./pinDetail.module.css";
 
 function PinDetailComponent() {
