+++ frontend/src/pages/ImageDetailPage.jsx
@@ -1,8 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import ShopItem from "./ImageDetailPage/ShopItem";
-import Breadcrumb from "./ImageDetailPage/Breadcrumb";
+
 import styles from "./Detail.module.css";
+import Breadcrumb from "./ImageDetailPage/Breadcrumb";
+import ShopItem from "./ImageDetailPage/ShopItem";
 
 function ImageDetailPage() {
   return (
