+++ frontend/src/pages/ImageDetailPage/CommentSection.jsx
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
 import { FaSmile, FaImage, FaGift, FaArrowRight } from "react-icons/fa";
+
 import styles from "./CommentSection.module.css";
 
 function CommentSection() {
