+++ package.json
@@ -42,6 +42,6 @@
     "globals": "^15.15.0",
     "lint-staged": "^15.5.1",
     "prettier": "^3.2.5",
-    "vite": "^5.1.4"
+    "vite": "^6.3.2"
   }
 }
